refactor(cli): rename `logs` to `logs-monitor` (#324)
We are about to move this module into a different package, and the
target package is gitignoring any directory called `logs`. We could just
fix the gitignore, but the module name is very inacurate anyway.
Therefore we are renaming it.

Also moves `rwlock` from `api/utils` to `api`. This was an oversight and
doesn't warrant a separate PR.

Finally moves some test files into their correct corresponding
directories.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license
