+++ packages/@aws-cdk/toolkit-lib/lib/api/cloud-assembly/index.ts
@@ -1,4 +1,4 @@
-export * from '../../api/shared-public';
+export { StackSelectionStrategy, StackSelector } from '../../api/shared-public';
 export * from './source-builder';
 export * from './types';
