+++ packages/@aws-cdk/toolkit-lib/lib/api/shared-private.ts
@@ -2,3 +2,4 @@
 
 export * from '../../../tmp-toolkit-helpers/src/api/io/private';
 export * from '../../../tmp-toolkit-helpers/src/private';
+export * from '../../../tmp-toolkit-helpers/src/api';
