+++ packages/@aws-cdk/toolkit-lib/lib/api/shared-public.ts
@@ -1,5 +1,21 @@
 /* eslint-disable import/no-restricted-paths */
 
-export * from '../../../tmp-toolkit-helpers/src/api';
+export {
+  ToolkitError,
+  AuthenticationError,
+  AssemblyError,
+  ExpandStackSelection,
+  StackSelectionStrategy,
+} from '../../../tmp-toolkit-helpers/src/api';
+
+export type {
+  IoMessageLevel,
+  IoMessageCode,
+  IIoHost,
+  IoMessage,
+  IoRequest,
+  ToolkitAction,
+  StackSelector,
+} from '../../../tmp-toolkit-helpers/src/api';
 
 export * from '../../../tmp-toolkit-helpers/src/payloads';
