+++ packages/@aws-cdk/toolkit-lib/lib/toolkit/index.ts
@@ -1,3 +1,2 @@
 export * from './toolkit';
 export * from './non-interactive-io-host';
-export * from '../api/shared-public';
