+++ packages/@aws-cdk/toolkit-lib/lib/toolkit/toolkit.ts
@@ -20,7 +20,7 @@ import type { WatchOptions } from '../actions/watch';
 import { patternsArrayForWatch } from '../actions/watch/private';
 import { type SdkConfig } from '../api/aws-auth';
 import type { SuccessfulDeployStackResult, StackCollection, Concurrency, AssetBuildNode, AssetPublishNode, StackNode } from '../api/aws-cdk';
-import { DEFAULT_TOOLKIT_STACK_NAME, Bootstrapper, SdkProvider, Deployments, HotswapMode, ResourceMigrator, tagsForStack, WorkGraphBuilder, CloudWatchLogEventMonitor, findCloudWatchLogGroups } from '../api/aws-cdk';
+import { DEFAULT_TOOLKIT_STACK_NAME, Bootstrapper, SdkProvider, Deployments, ResourceMigrator, tagsForStack, WorkGraphBuilder, CloudWatchLogEventMonitor, findCloudWatchLogGroups } from '../api/aws-cdk';
 import type { ICloudAssemblySource } from '../api/cloud-assembly';
 import { StackSelectionStrategy } from '../api/cloud-assembly';
 import type { StackAssembly } from '../api/cloud-assembly/private';
@@ -28,9 +28,8 @@ import { ALL_STACKS, CloudAssemblySourceBuilder, IdentityCloudAssemblySource } f
 import type { IIoHost, IoMessageLevel } from '../api/io';
 import { IO, SPAN, asSdkLogger, withoutColor, withoutEmojis, withTrimmedWhitespace } from '../api/io/private';
 import type { IoHelper } from '../api/shared-private';
-import { asIoHelper } from '../api/shared-private';
-import type { AssemblyData, StackDetails, ToolkitAction } from '../api/shared-public';
-import { DiffFormatter, RequireApproval, ToolkitError } from '../api/shared-public';
+import { asIoHelper, DiffFormatter, RequireApproval, ToolkitError } from '../api/shared-private';
+import type { ToolkitAction, AssemblyData, StackDetails } from '../api/shared-public';
 import { obscureTemplate, serializeStructure, validateSnsTopicArn, formatTime, formatErrorMessage } from '../private/util';
 import { pLimit } from '../util/concurrency';
 import { promiseWithResolvers } from '../util/promises';
