+++ packages/aws-cdk/lib/api/logs-monitor/find-cloudwatch-logs.ts
