+++ packages/aws-cdk/lib/api/logs-monitor/index.ts
