+++ packages/aws-cdk/lib/api/logs-monitor/logs-monitor.ts
