+++ packages/aws-cdk/lib/cli/cli.ts
@@ -26,9 +26,9 @@ import { Deployments } from '../api/deployments';
 import { HotswapMode } from '../api/hotswap/common';
 import { Notices } from '../api/notices';
 import { PluginHost } from '../api/plugin';
+import type { ILock } from '../api/rwlock';
 import type { Settings } from '../api/settings';
 import { ToolkitInfo } from '../api/toolkit-info';
-import type { ILock } from '../api/util/rwlock';
 import { contextHandler as context } from '../commands/context';
 import { docs } from '../commands/docs';
 import { doctor } from '../commands/doctor';
