+++ packages/aws-cdk/test/api/cloudformation/lazy-list-stack-resources.test.ts
@@ -1,8 +1,8 @@
 import 'aws-sdk-client-mock-jest';
 
 import { ListStackResourcesCommand } from '@aws-sdk/client-cloudformation';
-import { LazyListStackResources } from '../../lib/api/cloudformation';
-import { MockSdk, mockCloudFormationClient } from '../_helpers/mock-sdk';
+import { LazyListStackResources } from '../../../lib/api/cloudformation';
+import { MockSdk, mockCloudFormationClient } from '../../_helpers/mock-sdk';
 
 describe('Lazy ListStackResources', () => {
   test('correctly caches calls to the CloudFormation API', async () => {
