+++ packages/aws-cdk/test/api/cloudformation/lazy-lookup-export.test.ts
@@ -1,6 +1,6 @@
 import { ListExportsCommand } from '@aws-sdk/client-cloudformation';
-import { LazyLookupExport } from '../../lib/api/cloudformation';
-import { MockSdk, mockCloudFormationClient, restoreSdkMocksToDefault } from '../_helpers/mock-sdk';
+import { LazyLookupExport } from '../../../lib/api/cloudformation';
+import { MockSdk, mockCloudFormationClient, restoreSdkMocksToDefault } from '../../_helpers/mock-sdk';
 
 describe('LazyLookupExport', () => {
   const mockSdk = new MockSdk();
